enhanced_requirements.txt
# Enhanced Defense Intelligence ML Requirements
# Multi-platform GPU + Database Integration + Real Datasets

# Deep Learning Frameworks
tensorflow==2.15.0
torch==2.1.1

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0

# Data Science
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning
scikit-learn==1.3.2

# Signal Processing
librosa==0.10.1

# Database Integration
pymongo==4.6.0
mysql-connector-python==8.2.0

# Data Sources Integration
kaggle==1.5.16
requests==2.31.0

# Utilities
tqdm==4.66.1
joblib==1.3.2
h5py==3.10.0

# Development
jupyter==1.0.0
ipykernel==6.27.1

# GPU Monitoring (optional)
nvidia-ml-py==12.535.133
